fix(style): implement general transaction layout
